requirements.txt
Django
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn
whitenoise
google-auth